documentHandlerService/requirements.txt
django
djangorestframework
packaging
drf-yasg
readerwriterlock
mock